clns: bugfix: unsigned long != uint64_t on some systems

+++ src/inet/networklayer/contract/clns/ClnsAddress.cc
@@ -75,23 +75,23 @@ ClnsAddress::ClnsAddress(std::string net)
         case 2:
             dots++;
             // area[0] = (unsigned char) (atoi(net.substr(0, 2).c_str()));
-            areaID += strtoul(net.substr(0, 2).c_str(), NULL, 16) << 16;
+            areaID += (uint64_t)(strtoul(net.substr(0, 2).c_str(), NULL, 16)) << 16;
             break;
         case 7:
-            areaID += strtoul(net.substr(3, 4).c_str(), NULL, 16);
+            areaID += (uint64_t)(strtoul(net.substr(3, 4).c_str(), NULL, 16));
             dots++;
             break;
         case 12:
             dots++;
-            systemID += strtoul(net.substr(8, 4).c_str(), NULL, 16) << 32;
+            systemID += (uint64_t)(strtoul(net.substr(8, 4).c_str(), NULL, 16)) << 32;
             break;
         case 17:
             dots++;
-            systemID += strtoul(net.substr(13, 4).c_str(), NULL, 16) << 16;
+            systemID += (uint64_t)(strtoul(net.substr(13, 4).c_str(), NULL, 16)) << 16;
             break;
         case 22:
             dots++;
-            systemID += strtoul(net.substr(18, 4).c_str(), NULL, 16);
+            systemID += (uint64_t)(strtoul(net.substr(18, 4).c_str(), NULL, 16));
             break;
         default:
             return;
